% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{applyTransforms}
\alias{applyTransforms}
\title{Apply list of transformations to image}
\usage{
applyTransforms(i, trs = getRotations(), simplify = TRUE)
}
\arguments{
\item{i}{image (3D array)}

\item{trs}{single transformation or list of transformations (see \code{link{getRotations}})}

\item{simplify}{logical, should image be returned in case if only one transformation was supplied}
}
\value{
list of transformed images
}
\description{
Apply list of transformations to image
}
