% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{bandwidth.nrd.not0}
\alias{bandwidth.nrd.not0}
\title{Identify optimal bandwidth via normal distribution}
\usage{
bandwidth.nrd.not0(x)
}
\arguments{
\item{x}{a data vector}
}
\value{

}
\description{
ensure that result is not zero: removes the most frequent value in x if bandwidth.nrd return zero
}
