% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{barplotWithText}
\alias{barplotWithText}
\title{Plots barplot with text around bar tops}
\usage{
barplotWithText(
  x,
  t = x,
  adj = c(0.5, 1.1),
  srt = 0,
  text.y = x,
  ylim = c(0, max(x)),
  ...
)
}
\arguments{
\item{x}{heights of bars}

\item{t}{text to be plotted (\code{x} by default)}

\item{adj}{relative text position (see \code{\link{text}})}

\item{srt}{text rotation (see \code{\link{text}})}

\item{text.y}{text position (\code{x} by default)}

\item{ylim}{ylim}

\item{...}{other barplot options}
}
\description{
Barplot wrapper
}
