% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{calcColSums}
\alias{calcColSums}
\title{Calculates row sums for matrix subsets}
\usage{
calcColSums(d, f, mean = FALSE)
}
\arguments{
\item{d}{matrix}

\item{f}{character vector, factor to split matrix columns (length should be equal to nrow(d))}

\item{mean}{logical, calculate mean instead of sum}
}
\value{
matrix with number of rows equal to nrow(d) and number of columns equal to number of unique(f)
}
\description{
Calculates row sums for matrix subsets
}
