% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{calcDistance2border}
\alias{calcDistance2border}
\title{Classifies visium spots according to its position relative to tissue slice border}
\usage{
calcDistance2border(rc, nj)
}
\arguments{
\item{rc, nj}{output of findTissueBorder}
}
\value{
augmented rc dataframe with spot coordinates. Following columns added:
nearest.border.inxs.graph - list of indexes (as specified by border.inx) of all nearestest border spots
dist2border.graph - distance to tissue border (in spots, by hex-graph of spots)
nearest.border.pos.graph - mean of nearest.border.inxs.graph

please ignore other coumns added
}
\description{
Classifies visium spots according to its position relative to tissue slice border
}
