% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.in.depth.R
\name{calcDistance2SpotSet}
\alias{calcDistance2SpotSet}
\title{Calculates distance from each spot to the spot set}
\usage{
calcDistance2SpotSet(v, spots, image.name = NULL)
}
\arguments{
\item{v}{Seurat visium object}

\item{spots}{set of spots (character, intereger, or logical index of the spots in v)}

\item{image.name}{image to use, specify if object contains more than one sample}
}
\value{
numeric vector with distances to the nearest spot from set (in interspot distances)
}
\description{
Finds distance to the closest spot from the set and returns it expressed in inter-spot distances
}
