% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{calcMeanCols}
\alias{calcMeanCols}
\title{Calculates row sums for matrix subsets}
\usage{
calcMeanCols(d, f, FUN = base::mean, verbose = FALSE)
}
\arguments{
\item{d}{matrix}

\item{f}{character vector, factor to split matrix columns (length should be equal to nrow(d))}

\item{FUN}{function to be applied to matrix slices (mean by default)}

\item{verbose}{}
}
\value{
matrix with number of rows equal to nrow(d) and number of columns equal to number of unique(f)
}
\description{
Calculates row sums for matrix subsets
}
