% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.in.depth.R
\name{calcSpotDistance}
\alias{calcSpotDistance}
\title{Calculates matrix of spatial distances between spot centers}
\usage{
calcSpotDistance(v, image.name = NULL)
}
\arguments{
\item{v}{Seurat visium object}

\item{image.name}{image to use, specify if object contains more than one sample}
}
\value{
distance matrix
}
\description{
and normolazes it by inter-spot distance
}
