% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{col2hex}
\alias{col2hex}
\title{Transforms colour to hex representation}
\usage{
col2hex(c, withAlpha = TRUE)
}
\arguments{
\item{c}{character vector with colours}

\item{withAlpha}{}
}
\description{
Transforms colour to hex representation
}
