% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{compareSets}
\alias{compareSets}
\title{Compare two lists of sets}
\usage{
compareSets(l1, l2 = l1, fun = "o")
}
\arguments{
\item{l1}{list of sets to be compared}

\item{l2}{list of sets to be compared (compare l1 with itself if l2 is not set)}

\item{fun}{function to calculate similarity or character. 'o' will result in intersect/min; 'j' in intersect/union}
}
\value{
matrix if length(l1),length(l2) size
}
\description{
return matrix of set similarities defined by fun.
}
\details{
Default is size of intersection divided by minimal size (overlap or Szymkiewicz–Simpson coefficient).
}
\examples{
compareSets(list(letters[1:6],1:5),list(letters[1:8],4:8,letters[4:10]),fun = 'j')
}
