% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{cropVisiumImage}
\alias{cropVisiumImage}
\title{Crop image in Seuart Visium object}
\usage{
cropVisiumImage(v)
}
\arguments{
\item{v}{Seurat object}
}
\value{
visium object with cropped image
}
\description{
Retain square image that contains all spots (listed in the object). Spot coordinates adjasted accordingly.
}
