% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.in.depth.R
\name{defineJunction}
\alias{defineJunction}
\title{Annotate spots on border between two features}
\usage{
defineJunction(
  v,
  ann.column,
  which,
  contactTo,
  image.name = NULL,
  dist.thr = 1.2
)
}
\arguments{
\item{v}{Seurat visium object}

\item{ann.column}{name of meta.data column that holds feature annotation}

\item{which}{name if feature border spots should belong to}

\item{contactTo}{name if feature border spots should contact with}

\item{image.name}{image to use, specify if object contains more than one sample}

\item{dist.thr}{maximal distance (in distances between spots) to consider spots as neighbors}
}
\value{
data.frame with two columns logical 'junction' that specifies whether spot belongs to junction and numerical 'dist2junction' expressed in inter-spot distances
}
\description{
find spots that belongs to one feature but contact with another
}
