% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{dir.create.r}
\alias{dir.create.r}
\title{Creates folder and its parents
if not exists}
\usage{
dir.create.r(d)
}
\arguments{
\item{d}{folder to be created}
}
\value{

}
\description{
Creates folder and its parents
if not exists
}
