% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{dotPlot}
\alias{dotPlot}
\title{Plots matrix as dotPlot}
\usage{
dotPlot(
  m,
  rfun = sqrt,
  colfun = function(x) num2col(x, c("white", "yellow", "violet", "black")),
  grid = TRUE,
  grid.lty = 2,
  grid.col = "gray",
  max.cex = 0.9,
  xlab = "",
  ylab = "",
  ylab.cex = 1,
  xlab.cex = 1,
  colColours = NULL,
  rowColours = NULL,
  scaleWM = TRUE,
  pch = 19,
  ...
)
}
\arguments{
\item{m}{numeric matrix to be shown}

\item{rfun}{function to calculate radius from matrix values. Use sqrt (default) to make area proportional to value}

\item{colfun}{function to transform values to colour gradient}

\item{grid}{logial, should grid be plotted}

\item{grid.lty}{lty of grid}

\item{grid.col}{line color of grid}

\item{max.cex}{max size of dots}

\item{xlab, ylab}{axis labels}

\item{ylab.cex}{magnification label for ylabs}

\item{colColours}{colour matrix to plot annotation for columns (matrix with nrow equal to ncol(m); each column of colColours is annotation)}

\item{rowColours}{colour matrix to plot annotation for row (matrix with nrow equal to nrow(m); each column of rowColours is annotation)}

\item{scaleWM}{logical, specifies wheter computed radiuses should be scaled into \link{0,1} interval}

\item{pch}{point character (19 by default)}

\item{...}{other parameters to be passed to plot function}
}
\value{

}
\description{
Shows values in matix by point size and color
}
