% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{first2Upper}
\alias{first2Upper}
\title{Capitalize first letter of input text}
\usage{
first2Upper(t)
}
\arguments{
\item{t}{character vector}
}
\value{
t with first letter of each item capitalized
}
\description{
Capitalize first letter of input text
}
