% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{getColoursByDistance}
\alias{getColoursByDistance}
\title{Maps distance matrix into colour space}
\usage{
getColoursByDistance(d, use3D = FALSE, orderBySim = FALSE)
}
\arguments{
\item{d}{distance matrix to be used with MDS}

\item{use3D}{logical, specifies whether whole \link{0,1}^3 space should be used. About equally-bright 2D space is used otherwise (default)}

\item{orderBySim}{logical, whether objects should be ordered by similarity. Preserves original order otherwise (default)}
}
\description{
use MDS (via cmdscale) to map objects into colour space
}
\examples{
d=1-cor(t(mtcars))
c=getColoursByDistance(d,F)
plot(cmdscale(d),col=c,pch=19)
}
