% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{getConsensus}
\alias{getConsensus}
\title{Calculate consensus matrix for list of clusterings}
\usage{
getConsensus(cs)
}
\arguments{
\item{cs}{list of clusterings (output of getNMFcl)}
}
\value{
matrix n_celltypes*n_celltypes, with number of times given celltype pair was assigned to same cluster
}
\description{
Calculate consensus matrix for list of clusterings
}
