% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{getLineWidth}
\alias{getLineWidth}
\title{Returns with of line in user coordinates}
\usage{
getLineWidth(axis)
}
\arguments{
\item{axis}{character, either 'x' or 'y'}
}
\value{
with of line in user coordinates
}
\description{
Returns with of line in user coordinates
}
\examples{
getLineWidth('x')
}
