% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{getMaxInxByFirstDD}
\alias{getMaxInxByFirstDD}
\title{Find best match between image and RNA counts
to deal with image swaps/rotations}
\usage{
getMaxInxByFirstDD(m)
}
\arguments{
\item{m}{}
}
\value{
data.frame with best mtches
}
\description{
Find best match between image and RNA counts
to deal with image swaps/rotations
}
