% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{getMeanSpotColor}
\alias{getMeanSpotColor}
\title{Extracts mean color of spots}
\usage{
getMeanSpotColor(vis, scalefactors, image.name = NULL)
}
\arguments{
\item{vis}{Seurat object}

\item{scalefactors}{list of scale factors (see Details)}
}
\value{
matrix with three coumns (red, green, blue) and number of rows equal to the number of spots in vis object
}
\description{
Extracts mean color of spots
}
\details{
Seurat version 4.1.0 for stores wrong value for spot size in .
So one needs to load scale factors themself using jsonlite::read_json('/spatial/scalefactors_json.json')).
Output of read_json can be used as scalefactors parameter.
}
