% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{getMeanSpotColor1}
\alias{getMeanSpotColor1}
\title{Extracts mean color of given spot}
\usage{
getMeanSpotColor1(i, x, y, r)
}
\arguments{
\item{i}{image (3d array)}

\item{x, y}{coordinates of center of the spot}

\item{r}{radius of the spot}
}
\value{
numeric vector with three values (red, green, and blue)
}
\description{
Extracts mean color of given spot
}
