% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{getNMFcl}
\alias{getNMFcl}
\title{Get clusterings from list of nmfs}
\usage{
getNMFcl(nmfs, normF)
}
\arguments{
\item{nmfs}{list of nmf outputs}

\item{normF}{normalization function (output of getNMFNormFs)}
}
\value{
list of clusterings
}
\description{
Get clusterings from list of nmfs
}
