% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{getReadCoverage}
\alias{getReadCoverage}
\title{Extract read coverage from bam files}
\usage{
getReadCoverage(bams, chr, start, end, strand = NA)
}
\arguments{
\item{bams}{character vector with paths to bam files}

\item{chr}{contig name}

\item{start, end}{coordinates of region}

\item{strand}{strand, NA for unstranded (default)}
}
\value{
list with three items: x (chr coordinates); cov - number of reads mapped to chr position; juncs - data.frame with introns
}
\description{
Extract read coverage from bam files
}
