% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{getRotations}
\alias{getRotations}
\title{Return list of 8 possible mirroring/rotations function}
\usage{
getRotations()
}
\value{
list of functions
}
\description{
Return list of 8 possible mirroring/rotations function
}
