% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{getSilhouetteStat}
\alias{getSilhouetteStat}
\title{Assesses clusterings against distance matrix}
\usage{
getSilhouetteStat(cls, d)
}
\arguments{
\item{cls}{list of clusterings}

\item{d}{distance matrix (for example 1 - coincedence frequency)}
}
\value{
data.frame with silhouette summaries
}
\description{
Assesses clusterings against distance matrix
}
