% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{hist2D}
\alias{hist2D}
\title{2D density plot}
\usage{
hist2D(
  x,
  y,
  xbins = 100,
  ybins = 100,
  cols = c("white", "gray", "blue", "orange", "red"),
  zfun = identity,
  leg.title = "",
  num.leg.tic = NULL,
  legend = TRUE,
  trimZq = 0,
  xlim = NULL,
  ylim = NULL,
  new = TRUE,
  xlab = deparse(substitute(x)),
  ylab = deparse(substitute(y)),
  ...
)
}
\arguments{
\item{x, y}{point coordinates}

\item{xbins, ybins}{number of bins or bin borders. Everything outside of bins are removed}

\item{cols}{colours to form gradient}

\item{zfun}{transformation of frequency (try \code{\link{log1p}}, \code{\link{sqrt}}, etc). Identity by default}

\item{leg.title}{legend title}

\item{num.leg.tic}{desired number of tics in the legend}

\item{legend}{logical, should legend be plotted}

\item{trimZq}{frequencies outside of quantile(trimZq,1-trimZq) are set to the corresponding (closest) quantiles. Default is 0 (not trimming). See \code{\link{trimQ}}.}

\item{xlim, ylim, xlab, ylab}{graphical parameters}

\item{new}{create new plot (default) or add to existing}

\item{...}{other arguments for plot function}
}
\value{
list bith bins and frequencies (invisible)
}
\description{
Replacement of scatterplot to be used when number of points is to high and/or they highly overlap
}
