% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{isColors}
\alias{isColors}
\title{Tests whether value is colour}
\usage{
isColors(x)
}
\arguments{
\item{x}{vector to be tested}
}
\value{
logical vector
}
\description{
Tests whether value is colour
}
