% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{loadVisiumFrom10x}
\alias{loadVisiumFrom10x}
\title{Download Visium dataset from 10x website}
\usage{
loadVisiumFrom10x(url, sample.name, outdir)
}
\arguments{
\item{url}{url of the dataset (whithot dataset name, see example)}

\item{sample.name}{name of the dataset}

\item{outdir}{folder to save dataset (function will create subfolder named by \code{sample.name})}
}
\description{
Download Visium dataset from 10x website
}
\examples{
loadVisiumFrom10x("https://cf.10xgenomics.com/samples/spatial-exp/1.3.0","Visium_Mouse_Olfactory_Bulb",outdir="data")
}
