% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.in.depth.R
\name{makeDistFeatureSampleTable}
\alias{makeDistFeatureSampleTable}
\title{Summarises feature expression along specified factor and sample}
\usage{
makeDistFeatureSampleTable(
  dist,
  sample,
  data,
  filter = NULL,
  FUN = mean,
  min.spots = 5,
  ncores = 1,
  per.spot.norm = TRUE
)
}
\arguments{
\item{dist}{factor to summarise data along. Normally it is binnarised distance to something.}

\item{sample}{character vector specifying sample}

\item{data}{numeric matrix (probably sparse) with data to be summarised.}

\item{filter}{spot filter, the variable to be used to subset spots (will be applied to dist, sample, and data). No filtering applied if NULL (default)}

\item{FUN}{function to summarise the data (mean by default)}

\item{min.spots}{minimal namber of spots for statistics to be calculated (5 by default)}

\item{ncores}{number of cores to be used}

\item{per.spot.norm}{whether to perform per-spot normalization (makes sence for celltype abundancies)}
}
\value{
3d (distance/feature/sample) numeric matrix with summarized data
}
\description{
summary statistics will be calculated for each dist-sample-feature combination
}
