% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{mergePNG2PFD}
\alias{mergePNG2PFD}
\title{Merge multiple png files into pdf}
\usage{
mergePNG2PFD(dir = NULL, fls = NULL, pdfout, ...)
}
\arguments{
\item{dir}{path to the folder with pngfiles (used if \code{fls} is NULL)}

\item{fls}{character vector of paths to png files}

\item{pdfout}{name of output pdf}

\item{...}{other parameters for \code{\link{pdf}} function}
}
\value{

}
\description{
Merge multiple png files into pdf
}
