% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.spots.R
\name{mergeSpots}
\alias{mergeSpots}
\title{Merge grouped spots}
\usage{
mergeSpots(v, gr, image.name = NULL)
}
\arguments{
\item{v}{Seurat object}

\item{gr}{output of getCenters}
}
\value{
Seurat object with merged spots
}
\description{
Merge grouped spots
}
\examples{
c2 = getCenters(v2,to.merge = v2$nCount_Spatial<500)
v2m = mergeSpots(v2,c2)
v2m@meta.data[1:20,]
plotVisium(v2m,v2m$nCount_Spatial,zfun = log1p,cex=v2m$cex)
}
