% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{my.make.unique}
\alias{my.make.unique}
\title{Make Character Strings Unique}
\usage{
my.make.unique(x, sep = ".")
}
\arguments{
\item{x}{a character vector}

\item{sep}{a character string used to separate a duplicate name from its sequence number.}
}
\value{
A character vector of same length as names with duplicates changed
}
\description{
Makes the elements of a character vector unique by appending sequence numbers to duplicates.
The difference from make.unique is that my.make.unique adds replicate number to all items, not to duplicates only.
}
