% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{myLoad10X_Spatial}
\alias{myLoad10X_Spatial}
\title{Load spaceranger results}
\usage{
myLoad10X_Spatial(data.dir, filter.matrix = TRUE, ens_id = TRUE, ...)
}
\arguments{
\item{data.dir}{path to the folder with h5 files}

\item{filter.matrix}{logical, specifies whether only tissue spots should be loaded}

\item{ens_id}{logical, specifies whetehr Ensambl IDs should be used to identify genes (instead of gene names)}

\item{...}{other parameters to \code{\link[Seurat]{Load10X_Spatial}}}
}
\value{
Seurat object
}
\description{
Wrapper for \code{\link[Seurat]{Load10X_Spatial}}
}
