% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{myLoadH5AD_Spatial}
\alias{myLoadH5AD_Spatial}
\title{Loads scanpy h5ad file with Visium data into Seurat object}
\usage{
myLoadH5AD_Spatial(filename)
}
\arguments{
\item{filename}{character, path to h5ad file}
}
\value{
Seurat object
}
\description{
Loads scanpy h5ad file with Visium data into Seurat object
}
