% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{myLoadH5AD_Spatials}
\alias{myLoadH5AD_Spatials}
\title{Loads scanpy h5ad file with multiple Visium data into list of Seurat object}
\usage{
myLoadH5AD_Spatials(filename, library_id_field = "library_id")
}
\arguments{
\item{filename}{character, path to h5ad file}

\item{library_id_field}{name of adata.obs that specifies visium library. 'library_id' by default}
}
\value{
list of Seurat object
}
\description{
Loads scanpy h5ad file with multiple Visium data into list of Seurat object
}
