% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmf.R
\name{nmfGetBest}
\alias{nmfGetBest}
\title{Finds best nmf run}
\usage{
nmfGetBest(nmfs, normF)
}
\arguments{
\item{nmfs}{list of nmf results}

\item{normF}{normalization function (output of getNMFNormFs)}
}
\value{
list containing:
best nmf
normalized basis and coedicients (these are in addition per celltype sum-normalized)
celltype clustering
consensus matrix
Silhouette statistics
intex of best run
}
\description{
by 1) its agreement (mean Silhouette) with consensus matrix calculated based on all runs 2) deviance
}
\details{
the one that agree best with coincedence matrix
}
