% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{num2col}
\alias{num2col}
\title{Transforms numbers to colours}
\usage{
num2col(
  d,
  col = c("blue", "cyan", "gray", "orange", "red"),
  minx = min(d, na.rm = TRUE),
  maxx = max(d, na.rm = TRUE)
)
}
\arguments{
\item{d}{numbers to be transformed}

\item{col}{colours to form gradient (blue-red by default)}

\item{minx, maxx}{range of x values (min(x),max(x) by default)}
}
\value{
vector of colours (same length as d)
}
\description{
Wrapper function for colorRamp. Transforms numbers to the gradient specified by col parameter
}
