% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{number2bin}
\alias{number2bin}
\title{Splits numetic vector into equally sized bins}
\usage{
number2bin(v, n)
}
\arguments{
\item{v}{vector to be splitted}

\item{n}{number of bins}
}
\value{
numeric vector win bin number (same length as \code{v})
}
\description{
Splits numetic vector into equally sized bins
}
