% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{overlayColours}
\alias{overlayColours}
\title{Calculate per-row mean colour according to opacity}
\usage{
overlayColours(c, reorderByOpacity = FALSE)
}
\arguments{
\item{c}{text matrix with colours to be overlayed per row}

\item{reorderByOpacity}{logical, should colours be ordered by increased opacity prior to summing}
}
\value{

}
\description{
Calculate per-row mean colour according to opacity
}
