% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{overlayTwoColours}
\alias{overlayTwoColours}
\title{Sums two colours with alpha channel}
\usage{
overlayTwoColours(a, b)
}
\arguments{
\item{a}{top colour (numeric vector of length four)}

\item{b}{bottom colour (numeric vector of length four)}
}
\value{
vector of length four with resultant colour
}
\description{
Sums two colours with alpha channel
}
