% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{plotArc}
\alias{plotArc}
\title{Plots parabolic arc}
\usage{
plotArc(from, to, top, n = 100, y.base = 0, ...)
}
\arguments{
\item{from, to}{x coordinates of arc}

\item{top}{highest point of arc}

\item{n}{number of points}

\item{y.base}{bottom coordinate of arc}

\item{...}{other parameters of lines functoin}
}
\description{
Plots parabolic arc
}
