% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{plotColorLegend}
\alias{plotColorLegend}
\title{Gradient legend}
\usage{
plotColorLegend(
  x0,
  x1,
  y0,
  y1,
  col,
  at,
  legend,
  title = NULL,
  title.adj = c(0, -0.5)
)
}
\arguments{
\item{x0, x1, y0, y1}{rectange coordinates of the legend in nfc coordinates}

\item{col}{vector of colors (full gradient)}

\item{at}{indexes (of col vector) to place tics}

\item{legend}{tics labels}

\item{title}{legend title}

\item{title.adj}{legend title adj parameter to be passed to text function}
}
\description{
Gradient legend
}
\examples{
plot(1)
plotColorLegend(0.4,0.5,0.8,0.3,getPal(n=100),0:10*10,1:10)
}
