% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{plotColorLegend2}
\alias{plotColorLegend2}
\title{Gradient legend}
\usage{
plotColorLegend2(
  x0 = NULL,
  x1 = NULL,
  y0 = NULL,
  y1 = NULL,
  zlim,
  fullzlim = zlim,
  zfun = identity,
  z2col = num2col,
  N = 100,
  ntic = 5,
  leg = NULL,
  title = NULL,
  title.adj = c(0, -0.5),
  horizontal = FALSE
)
}
\arguments{
\item{x0, x1, y0, y1}{rectange coordinates of the legend in nfc coordinates}

\item{zlim}{numeric vector with two items. Range to show, trimmed to fullzlim if wider.}

\item{fullzlim}{numeric vector with two items. Full range to be considered.}

\item{zfun}{transformation for value (gradient will be drawn along transformed value). Identity, log1p, sqrt, ets.}

\item{z2col}{function to transform numbers to colors}

\item{N}{number of steps in the gradient}

\item{ntic}{desired number of tics}

\item{leg}{tics values, if NULL (default) estimated automatically}

\item{title}{legend title}

\item{title.adj}{legend title adj parameter to be passed to text function}
}
\description{
Wrapper for plotColorLegend to plot legend for given range
}
