% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.in.depth.R
\name{plotConditionsProfiles}
\alias{plotConditionsProfiles}
\title{Plot feature profile along distance for multiple conditions}
\usage{
plotConditionsProfiles(
  m,
  feature,
  conditions,
  cols = NULL,
  ltys = 1,
  sd.mult = 2,
  legend. = TRUE,
  ylim = NULL,
  xlim = NULL,
  scaleY = FALSE,
  area.opacity = 0.2,
  lwd = 2,
  xlab = "Distance (spots)",
  ylab = "Abundance",
  main = "",
  cilim = c(-Inf, Inf)
)
}
\arguments{
\item{m}{3d (distance/feature/sample) numeric matrix with summarized data (output of makeDistFeatureSampleTable)}

\item{feature}{single character, name of feature to be plotted (from second dimension of m)}

\item{conditions}{vector specifying conditions of all samples (along third dimension of m)}

\item{cols}{named (by conditions) color vector. Only conditions mentioned here will be shown. Colours are aoutogenerated by char2col if null (default)}

\item{sd.mult}{width of CI shown by shade expressed in standart deviations of mean}

\item{legend.}{logical, whether logend should be plotted, or list with parameters to be passed to legend function.}

\item{ylim}{ylim to be set, set by data if NULL (default)}

\item{scaleY}{whether all features should be scaled by its max value (to bring all features to the same scale)}

\item{area.opacity}{opacity of CI shade}

\item{lwd}{line width}

\item{xlab, ylab, main}{plot annotation}

\item{cilim}{numerical vector with two values, gives lower and apper values to truncate CI. NULL (to truncation) by default.}

\item{...}{other parameters to be passed to plot}
}
\description{
Plot feature profile along distance for multiple conditions
}
