% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.in.depth.R
\name{plotFeatureProfiles}
\alias{plotFeatureProfiles}
\title{Plots features profiles}
\usage{
plotFeatureProfiles(
  m,
  features,
  cols = NULL,
  sd.mult = 2,
  legend. = TRUE,
  ylim = NULL,
  scaleY = TRUE,
  area.opacity = 0.2,
  lwd = 2,
  xlab = "Distance (spots)",
  ylab = "Relative abundance",
  main = "",
  ...
)
}
\arguments{
\item{m}{3d (distance/feature/sample) numeric matrix with summarized data (output of makeDistFeatureSampleTable)}

\item{features}{features (indexes along third dimension of m)}

\item{cols}{names (by feature names) colour vector}

\item{sd.mult}{width of CI shown by shade expressed in standart deviations of mean}

\item{legend.}{logical, whether logend should be plotted, or list with parameters to be passed to legend function.}

\item{ylim}{ylim to be set, set by data if NULL (default)}

\item{scaleY}{whether all features should be scaled by its max value (to bring all features to the same scale)}

\item{area.opacity}{opacity of CI shade}

\item{lwd}{line width}

\item{xlab, ylab, main}{plot annotation}

\item{...}{other parameters to be passed to plot}
}
\description{
along distance. Shows mean abundance as line and CI as shade.
}
