% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{plotLine}
\alias{plotLine}
\title{Plots scatterplot with regression line}
\usage{
plotLine(
  x,
  y = NULL,
  cor.method = "pearson",
  line.col = "red",
  line.lwd = 1,
  plot.ci = FALSE,
  ci.transparency = 0.3,
  line.on.top = TRUE,
  new = TRUE,
  ...
)
}
\arguments{
\item{x, y}{coordinates}

\item{cor.method}{pearson (default) or spearman}

\item{line.col}{colour of line}

\item{line.lwd}{line width}

\item{plot.ci}{logical, should CI be plotted}

\item{ci.transparency}{alpha of CI area}

\item{line.on.top}{logical, should line be on the top of points}

\item{new}{logical, create new plot (default) or add to existing}

\item{...}{other arguments for plot function}
}
\value{

}
\description{
Wrapper for pllot function
}
