% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{plotNMFCons}
\alias{plotNMFCons}
\title{Plots NMF clustering, silhouette and consensuse clustering}
\usage{
plotNMFCons(
  coefs,
  cons,
  clcols = NULL,
  max.cex = 4.5/14 * 8,
  colfun = function(x) num2col(x, c("blue", "gray", "orange", "violet", "black"), minx =
    0, maxx = 1),
  title = "",
  ylab.cex = 1,
  xlab.cex = 1
)
}
\arguments{
\item{coefs}{coef matrix from NMF}

\item{cons}{conensus matrix}

\item{clcols}{colors for clusters}

\item{max.cex}{dot size for dotPlot}

\item{colfun}{color function to be used in dotPlot}

\item{title}{title for whole plot}

\item{ylab.cex, xlab.cex}{sizes of axis labels}
}
\value{
list of clusters, colors and silhouette
}
\description{
Plots NMF clustering, silhouette and consensuse clustering
}
