% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{plotPanelLetter}
\alias{plotPanelLetter}
\title{Add panel label}
\usage{
plotPanelLetter(l, cex = 1.2, adj = c(0, 1.1), ...)
}
\arguments{
\item{l}{letter to be added}

\item{cex}{letter size}

\item{adj}{see adj parameter of \code{\link{text}}}

\item{...}{other parameters for text function}
}
\description{
Adds letter into top left corner of the plot
}
