% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{plotPointDensity}
\alias{plotPointDensity}
\title{Scatterplot wih point density shown by colour}
\usage{
plotPointDensity(x, y, pch = 16, bty = "n", log = "", ...)
}
\arguments{
\item{x, y}{point coordinates}

\item{pch, bty, log}{parameters of plot function}

\item{...}{other parameters to be passed to plot}
}
\description{
Scatterplot wih point density shown by colour
}
