% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{plotReadCov}
\alias{plotReadCov}
\title{Title}
\usage{
plotReadCov(
  r,
  min.junc.cov = 0,
  plot.junc.only.within = FALSE,
  ylim = NULL,
  xlim = range(r$x),
  reverse = FALSE,
  junc.col = "red",
  junc.lwd = 3,
  ...
)
}
\arguments{
\item{cov}{list with three elements: x (coordinates), cov (coverage) and juncs (start,end,score)}
}
\value{

}
\description{
Title
}
