% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.in.depth.R
\name{plotTD.HM}
\alias{plotTD.HM}
\title{Plots results of testTDConditions}
\usage{
plotTD.HM(
  comp,
  cond.titles = c("cond1", "cond2"),
  order = NULL,
  l2fc.zlim = NULL,
  feature.class = NULL,
  log = TRUE,
  col = hcl.colors(100, "YlOrRd", rev = TRUE),
  fdr.thrs = c(`*` = 0.05, . = 0.2),
  feature.names.map = NULL
)
}
\arguments{
\item{comp}{list, output of testTDConditions}

\item{cond.titles}{condition titles to be plot on top of heatmaps}

\item{order}{feature order, NULL (default) for no reordering}

\item{l2fc.zlim}{range to trim log fold change}

\item{feature.class}{list (or matrix of columns) with feature annotations to be shown by color}

\item{fdr.thrs}{named numerical list used to show levels of significance on log fold change plot, default is c('*'=0.05,'.'=0.2)}

\item{feature.names.map}{named vector that maps feature ids used in comp to display names}

\item{cols}{cols to be used for condition heatmaps}
}
\description{
as three heatmaps: two shows mean per-condition spatial feature distribution and one show log fold change
}
