% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{plotVisium}
\alias{plotVisium}
\title{Plot Visium sample}
\usage{
plotVisium(
  v,
  z = NA,
  cex = 1,
  type = "img",
  border = NA,
  z2col = num2col,
  plot.legend = TRUE,
  zlim = NULL,
  zfun = identity,
  spot.filter = NULL,
  pch = 16,
  num.leg.tic = NULL,
  label.clusters = FALSE,
  legend.args = list(),
  randomize.points = FALSE,
  order.points.by.z = FALSE,
  xaxt = "n",
  yaxt = "n",
  cluster.lab.adj = c(0.5, 0.5),
  cluster.lab.cex = 1,
  cluster.lab.font = 1,
  cluster.lab2col = NULL,
  show.cluster.sizes = FALSE,
  ...
)
}
\arguments{
\item{v}{Seurat object or data.frame with two columns (x and y coordinates)}

\item{z}{z coordinate. Either numeric or categorical (factor/character)}

\item{cex}{relative size of symbols}

\item{type}{character, one of img, hex, rect, or xy, see Details.}

\item{border}{colour of symbol borders}

\item{z2col}{function to transform z to colous (if z is numeric) or named vector of colours if z is character (names are levels of \code{z})}

\item{plot.legend}{logical}

\item{zlim}{numerical vector with two items, used to trim z}

\item{zfun}{function, transformation (log1p, sqrt, ect) of z (used only if z is numerical). Identity by default.}

\item{spot.filter}{logical vector, specifies which spots should be plotted}

\item{pch}{point type to be used with type='xy'}

\item{num.leg.tic}{numerical, desired number of tics in gradient legend}

\item{label.clusters}{label cluster on top of spots. Either logical, or character, if latter should provide names of each spot, it this case could be different from \code{z}}

\item{legend.args}{list of arguments to be passed to legend function (for categorical z), only title is used for numerical z.}

\item{randomize.points}{logical, should points be randomazed (makes sense if type is 'xy' and if spots overlap and ordered by z)}

\item{order.points.by.z}{logical, make sense to show values with high z on the top if type is 'xy'  (doublets for instance)}

\item{xaxt, yaxt}{type of axis (see \code{par}), no axis are drown by default}

\item{cluster.lab.adj}{adj parameterfor cluster labels (see  \code{text})}

\item{cluster.lab.cex}{size of cluster labels}

\item{cluster.lab.font}{font of cluster labels}

\item{cluster.lab2col}{named vector that maps cluster names to colors to be used as text colors. All labels are black if NULL (default). Lables not mentioned in here will be shown in black as well.}

\item{show.cluster.sizes}{logical, specifies whether number of cell per cluster should be shown in the legend}

\item{...}{other arguments to be passed to graphical functions (see Details)}
}
\value{

}
\description{
Main function to plot both histological and dimention reduction plots
}
\details{
Plots spots on top of H&E image, if type is 'img' (see \code{\link{plotVisiumImg}} for additional parameters),
plots hexagonal representation if type is 'hex' (see \code{\link{plotVisiumHex}}). Type 'xy' forces dimention redutcion plot (by \code{plot}),
coordinates either directly rpovided in \code{v} (as two-column data.frame) or extracted from umap slot of \code{v},
in the later case function attempts to use seurat_clusters if \code{z} is not specified
}
