% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.graphics.R
\name{plotVisiumHex}
\alias{plotVisiumHex}
\title{Plot Visium as honeycombs}
\usage{
plotVisiumHex(
  xy,
  cex = 1,
  col = "red",
  border = col,
  xlab = "Cols",
  ylab = "Rows",
  xlim = c(min(xy$col) - 1, max(xy$col) + 1),
  ylim = c(max(xy$row) + 1, min(xy$row) - 1),
  hexstep = 0.25,
  transpose = FALSE,
  ...
)
}
\arguments{
\item{xy}{seu@images[\link{.}]@coordinates}

\item{cex}{spot size}

\item{col}{spot color}

\item{border}{color of spot border (by default)}

\item{xlim, ylim, xlab, ylab}{parameters of \code{plot} function}

\item{hexstep}{size of spot hex, do not change unless you know what you are doing}

\item{transpose}{logical, whether to transpose the figure}

\item{...}{other parameters to be passed to \code{plot} function}
}
\value{
data.frame with spot coordinates
}
\description{
Normally this function shouldn't be called directly. Use plotVisium instead
}
