% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{plotVisiumImg}
\alias{plotVisiumImg}
\title{Plot Visium sample on top of H&E}
\usage{
plotVisiumImg(
  xy,
  img,
  scale.factor,
  spot.radius,
  cex = 1,
  col = "red",
  border = NA,
  spot.dist = NULL,
  img.alpha = 1,
  xlim = NULL,
  ylim = NULL,
  symmetric.lims = TRUE,
  xlab = "",
  ylab = "",
  pie.fracs = NULL,
  pie.cols = NULL,
  pie.min.frac = 0.05,
  he.img.width = NULL,
  he.grayscale = FALSE,
  ...
)
}
\arguments{
\item{xy}{seu@images$slice1@coordinates}

\item{img}{image (3D array)}

\item{scale.factor}{see @images$slice1@scale.factors$lowres}

\item{spot.radius}{see @images$slice1@spot.radius}

\item{cex}{spot size}

\item{col}{spot color}

\item{border}{color of spot border}

\item{spot.dist}{distance between spots (computed from xy by default)}

\item{img.alpha}{alpha level for H&E image}

\item{xlim, ylim, xlab, ylab}{parameters of \code{plot} function}

\item{symmetric.lims}{logical, specifies whether image crop should be square}

\item{pie.fracs}{if specified plots pies instead of simple cycles. Matrix with number of rows equal to the number of spots, and number of columns equal to pie pieces.}

\item{pie.cols}{colors to be used for pie pieces (ncol(pie.fracs) should be equal to length(pie.cols))}

\item{pie.min.frac}{all pieces with relative size less than \code{pie.min.frac} will be discared}

\item{he.img.width}{integer, defines width (in pixels) the H&E figure should be resized to. No resizing if NULL (default)}

\item{he.grayscale}{logical, specifies whether H&E image should be converted to grayscale}

\item{...}{other parameters to be passed to \code{plot} function}
}
\description{
Normally this function shouldn't be called directly. Use plotVisium instead
}
