% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{plotVisiumMultyColours}
\alias{plotVisiumMultyColours}
\title{Plot multiple numerical values (gene expression or cell abundancies) on H&E image}
\usage{
plotVisiumMultyColours(
  v,
  x,
  cols = NULL,
  log.pc = NA,
  scale.per.colour = TRUE,
  reorderByOpacity = FALSE,
  title.adj = c(0, -0.5),
  bg = "#FFFFFFFF",
  legend.ncol = 1,
  ...
)
}
\arguments{
\item{v}{Seurat object}

\item{x}{matrix with values (in columns) to be plotted}

\item{cols}{colour to be used for x columns}

\item{log.pc}{numeric, pseudocount to be added before log-transformation. No transformation is applied if NA.}

\item{scale.per.colour}{logical, specifies whether each colour should cover whole range (that is, should x be sclaed per column)}

\item{reorderByOpacity}{logical, specifes whether colours should be ordered by increasing opacity prior to summing}

\item{title.adj}{legend title adj (to be passed to text function)}

\item{bg}{color to use as spot background. NULL (default) for transparent background.}

\item{legend.ncol}{number of legend columns. Set to 0 to suppress legend plotting.}

\item{...}{other parameters to be passed to plotVisium}
}
\value{

}
\description{
Values are plotted as a sum of color gradients. Be carefull, result depends on the order of features.
}
