% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{recycle}
\alias{recycle}
\title{Recycle vector v along i}
\usage{
recycle(v, i)
}
\arguments{
\item{v}{vector to be recycled}

\item{i}{either number of values to be obtained or 1:number_of_values}
}
\value{
recycled vector
}
\description{
Recycle vector v along i
}
