% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{renameClustsBySize}
\alias{renameClustsBySize}
\title{Rename cluster by size}
\usage{
renameClustsBySize(x)
}
\arguments{
\item{x}{clustering, numeric vector with membership (as returned by cutree for instance)}
}
\value{
numeric cluster with new membership
}
\description{
Gives larger clusters smaller index
}
