% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{rotateVisium}
\alias{rotateVisium}
\title{Rotates Seurat visium slide}
\usage{
rotateVisium(v, n = 1, mirror = FALSE)
}
\arguments{
\item{v}{Seurat object}

\item{n}{number of counter-clockwise rotations}

\item{mirror}{logical, whether to slide should be flipped horisontally}
}
\value{
rotated Seurat object
}
\description{
Rotates Seurat visium slide
}
\examples{
vm = rotateVisium(v,n=0,m=T)
SpatialFeaturePlot(v,'lnCount_Spatial')+SpatialFeaturePlot(vm,'lnCount_Spatial')
}
