% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{scaleTo}
\alias{scaleTo}
\title{Scale numeric vector}
\usage{
scaleTo(
  x,
  from = 0,
  to = 1,
  minx = min(x, na.rm = TRUE),
  maxx = max(x, na.rm = TRUE),
  fraction = 1
)
}
\arguments{
\item{x}{numeric vector to be scaled}

\item{from, to}{target range}

\item{minx, maxx}{true min of "real x" (in case if given x is just subset of the real)}

\item{fraction}{fraction of target range to be used}
}
\description{
Scale numeric vector
}
