% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{scaleVisiumImage}
\alias{scaleVisiumImage}
\title{Resizes H&E image}
\usage{
scaleVisiumImage(v, wpx = 500, image.name = NULL)
}
\arguments{
\item{v}{Seurat visium image}

\item{wpx}{desired image width (pixels)}
}
\value{
modified Seurat object
}
\description{
Resizes H&E image
}
