% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{splitSub}
\alias{splitSub}
\title{Cuts part of delimited text
analogous to bash cut}
\usage{
splitSub(x, del, inx, fixed = TRUE, simplify = TRUE)
}
\arguments{
\item{x}{character vector to cut from}

\item{del}{delimiter}

\item{inx}{indexes of items to return}

\item{fixed}{logical, whether del is fixed (to be passed to strsplit)}

\item{simplify}{logical, whether output should be simplified (to be passed to sapply)}
}
\value{
character vector or
}
\description{
Cuts part of delimited text
analogous to bash cut
}
\examples{
splitSub(c('a,b','d,c'),',',2)
}
