% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{symbols.pie}
\alias{symbols.pie}
\title{Plots scatter pie}
\usage{
symbols.pie(x, y, r, d, cols, border = NA, alpha0 = 0)
}
\arguments{
\item{x, y}{positions of pies to plot}

\item{r}{radiuses of pies (recycled along x)}

\item{d}{matrix of pie piece sizes. Row sums are normolized to 1. Nrow(d) should be equal to lenght(x).}

\item{cols}{colours of pieces. Ncol(d) should be equal to length(cols)}

\item{border}{colour of pie borders, NA (default) for no borders.}

\item{alpha0}{start angle}
}
\description{
Plots set of pie charts in specified positions
}
\examples{
plot(1,t='n',xlim=c(0,5),ylim=c(0,10))
d = matrix(runif(6*4),ncol=4)
symbols.pie(c(1,2,3,1,2,3),c(1,1,1,1,1,1),1/2,d,1:4)
}
