% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue.in.depth.R
\name{testTDConditions}
\alias{testTDConditions}
\title{Compare features bewteeb two conditions}
\usage{
testTDConditions(m, f1, f2, test.fun = t.test)
}
\arguments{
\item{m}{3d (distance/feature/sample) numeric matrix with summarized data (output of makeDistFeatureSampleTable)}

\item{f1, f2}{indexes that specify samples belonging to first and second conditions}

\item{test.fun}{function to be used to test difference. Should return list of p.value item. t.test is default}
}
\value{
list of pv, fdr, m1, and m2 (condition means) distance/feature matricex
}
\description{
per distance bean
}
