% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{trimQ}
\alias{trimQ}
\title{Trim by quantiles}
\usage{
trimQ(x, q)
}
\arguments{
\item{x}{numeric vector}

\item{q}{quantile}
}
\value{
corrected x
}
\description{
Sets all values of x below qth quantile to the quantile. Sets all values of x above (1-q)th quantile to the quantile.
}
