% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{enhanceImage}
\alias{enhanceImage}
\title{Adjust image brightnes and contrast}
\usage{
enhanceImage(p, wb = FALSE, qs = NULL, trim01 = TRUE)
}
\arguments{
\item{p}{image (3d numeric array)}

\item{wb}{logical, specifies whether output image should be transformed to grayscale}

\item{qs}{quantiles to trim. Numerical vector with two items. Trims all values outside of specified quantile range.}

\item{trim01}{logical, specifies whether pixels with zero (black) and maximal (white) intensity should be trimed ahead of quantile trimming.}
}
\value{
image (3d numeric array)
}
\description{
Adjust image brightnes and contrast
}
