% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{imageWithText}
\alias{imageWithText}
\title{Plot heatmap with numbers}
\usage{
imageWithText(
  d,
  t = NULL,
  digits = 2,
  text.col = NULL,
  xaxlab = rownames(d),
  yaxlab = colnames(d),
  centerColors0 = FALSE,
  las = 2,
  text.adj = c(0.5, 0.5),
  ...
)
}
\arguments{
\item{d}{matrix to be plotted}

\item{t}{text to be ploted on top of image (rounds d by default)}

\item{digits}{number of digits retained under rounding of d}

\item{text.col}{color for text}

\item{xaxlab, yaxlab}{axis labels, dimnames(d) by default}

\item{centerColors0}{logical, specifies whether 0 should be considered as palette center (makes sense for correlation matrices)}

\item{las}{orientation of axis labels (see las in ?par)}

\item{text.adj}{adj parameter to be passed to text function}

\item{...}{other options to be supplied to image}
}
\description{
Wraps image functions
}
