% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.purpuse.utilities.R
\name{plotColorLegend2}
\alias{plotColorLegend2}
\title{Gradient legend}
\usage{
plotColorLegend2(
  x0,
  x1,
  y0,
  y1,
  fullzlim,
  zlim,
  zfun,
  z2col,
  N = 100,
  ntic = 5,
  leg = NULL,
  title = NULL
)
}
\arguments{
\item{x0, x1, y0, y1}{rectange coordinates of the legend in nfc coordinates}

\item{fullzlim}{numeric vector with two items. Full range}

\item{zlim}{numeric vector with two items. Range to show (should be within fullzlim)}

\item{zfun}{transformation for value (gradient will be drawn along transformed value). Identity, log1p, sqrt, ets.}

\item{z2col}{function to transform numbers to colors}

\item{N}{number of steps in the gradient}

\item{ntic}{desired number of tics}

\item{leg}{tics values, if NULL (default) estimated automatically}

\item{title}{legend title}
}
\value{

}
\description{
Wrapper for plotColorLegend to plot legend for given range
}
