% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visium.utilities.R
\name{plotVisiumMultyColours}
\alias{plotVisiumMultyColours}
\title{Plot multiple numerical values (gene expression or cell abundancies) on H&E image}
\usage{
plotVisiumMultyColours(
  v,
  z,
  cols = NULL,
  zfun = identity,
  scale.per.colour = TRUE,
  mode = "overlay",
  reorderByOpacity = FALSE,
  min.opacity = 0,
  title.adj = c(0, -0.5),
  bg = "#FFFFFFFF",
  legend.ncol = 1,
  ...
)
}
\arguments{
\item{v}{Seurat object}

\item{z}{matrix with values (in columns) to be plotted}

\item{cols}{colour to be used for x columns}

\item{zfun}{function to transform values in z}

\item{scale.per.colour}{logical, specifies whether each colour should cover whole range (that is, should x be sclaed per column)}

\item{mode}{'overlay' or 'mean'.}

\item{reorderByOpacity}{logical, specifes whether colours should be ordered by increasing opacity prior to summing}

\item{min.opacity}{minumal spot opacity for mean mode (from 0 to 255)}

\item{title.adj}{legend title adj (to be passed to text function)}

\item{bg}{color to use as spot background. NULL (default) for transparent background.}

\item{legend.ncol}{number of legend columns. Set to 0 to suppress legend plotting.}

\item{...}{other parameters to be passed to plotVisium}
}
\value{

}
\description{
Values are plotted as a sum of color gradients. Be carefull, in overlay mode results can be misleading for spots with two dominant features.
}
